Graceful exit when wrong file is fed to muxer.
Added null check to output_uri in gst_add_track to return if set_data_sink fails.
Modify mediamuxer to return an error if the media_format during add_track is not matching with media_format of the write_sample packet
Update mediamuxer test-suite to return an error if inappropriate A/V codecs were fed.
Hence, user is restricted to feed a container-file with same encoded-media as what he choose during 'o'

Change-Id: Id8f4fe3637564e9957084f77dab5f6d856b4b2f6
Signed-off-by: Mahesh Kondiparthi <kd.mahesh@samsung.com>